feat(streaming): add tools support
