--- src/anthropic/lib/streaming/beta/__init__.py
@@ -1,14 +1,0 @@
-from ._tools import (
-    ToolsBetaMessageStream as ToolsBetaMessageStream,
-    ToolsBetaMessageStreamT as ToolsBetaMessageStreamT,
-    AsyncToolsBetaMessageStream as AsyncToolsBetaMessageStream,
-    AsyncToolsBetaMessageStreamT as AsyncToolsBetaMessageStreamT,
-    ToolsBetaMessageStreamManager as ToolsBetaMessageStreamManager,
-    AsyncToolsBetaMessageStreamManager as AsyncToolsBetaMessageStreamManager,
-)
-from ._types import (
-    ToolsBetaInputJsonEvent as ToolsBetaInputJsonEvent,
-    ToolsBetaMessageStopEvent as ToolsBetaMessageStopEvent,
-    ToolsBetaMessageStreamEvent as ToolsBetaMessageStreamEvent,
-    ToolsBetaContentBlockStopEvent as ToolsBetaContentBlockStopEvent,
-)
