+++ src/anthropic/resources/messages.py
@@ -953,6 +953,8 @@ class Messages(SyncAPIResource):
         temperature: float | NotGiven = NOT_GIVEN,
         top_k: int | NotGiven = NOT_GIVEN,
         top_p: float | NotGiven = NOT_GIVEN,
+        tool_choice: message_create_params.ToolChoice | NotGiven = NOT_GIVEN,
+        tools: Iterable[ToolParam] | NotGiven = NOT_GIVEN,
         # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
         # The extra values given here take precedence over values defined on the client or passed to this method.
         extra_headers: Headers | None = None,
@@ -986,6 +988,8 @@ class Messages(SyncAPIResource):
         temperature: float | NotGiven = NOT_GIVEN,
         top_k: int | NotGiven = NOT_GIVEN,
         top_p: float | NotGiven = NOT_GIVEN,
+        tool_choice: message_create_params.ToolChoice | NotGiven = NOT_GIVEN,
+        tools: Iterable[ToolParam] | NotGiven = NOT_GIVEN,
         event_handler: type[MessageStreamT],
         # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
         # The extra values given here take precedence over values defined on the client or passed to this method.
@@ -1019,6 +1023,8 @@ class Messages(SyncAPIResource):
         temperature: float | NotGiven = NOT_GIVEN,
         top_k: int | NotGiven = NOT_GIVEN,
         top_p: float | NotGiven = NOT_GIVEN,
+        tool_choice: message_create_params.ToolChoice | NotGiven = NOT_GIVEN,
+        tools: Iterable[ToolParam] | NotGiven = NOT_GIVEN,
         event_handler: type[MessageStreamT] = MessageStream,  # type: ignore[assignment]
         # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
         # The extra values given here take precedence over values defined on the client or passed to this method.
@@ -1047,6 +1053,8 @@ class Messages(SyncAPIResource):
                     "temperature": temperature,
                     "top_k": top_k,
                     "top_p": top_p,
+                    "tools": tools,
+                    "tool_choice": tool_choice,
                     "stream": True,
                 },
                 message_create_params.MessageCreateParams,
@@ -1975,6 +1983,8 @@ class AsyncMessages(AsyncAPIResource):
         temperature: float | NotGiven = NOT_GIVEN,
         top_k: int | NotGiven = NOT_GIVEN,
         top_p: float | NotGiven = NOT_GIVEN,
+        tool_choice: message_create_params.ToolChoice | NotGiven = NOT_GIVEN,
+        tools: Iterable[ToolParam] | NotGiven = NOT_GIVEN,
         # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
         # The extra values given here take precedence over values defined on the client or passed to this method.
         extra_headers: Headers | None = None,
@@ -2008,6 +2018,8 @@ class AsyncMessages(AsyncAPIResource):
         temperature: float | NotGiven = NOT_GIVEN,
         top_k: int | NotGiven = NOT_GIVEN,
         top_p: float | NotGiven = NOT_GIVEN,
+        tool_choice: message_create_params.ToolChoice | NotGiven = NOT_GIVEN,
+        tools: Iterable[ToolParam] | NotGiven = NOT_GIVEN,
         event_handler: type[AsyncMessageStreamT],
         # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
         # The extra values given here take precedence over values defined on the client or passed to this method.
@@ -2041,6 +2053,8 @@ class AsyncMessages(AsyncAPIResource):
         temperature: float | NotGiven = NOT_GIVEN,
         top_k: int | NotGiven = NOT_GIVEN,
         top_p: float | NotGiven = NOT_GIVEN,
+        tool_choice: message_create_params.ToolChoice | NotGiven = NOT_GIVEN,
+        tools: Iterable[ToolParam] | NotGiven = NOT_GIVEN,
         event_handler: type[AsyncMessageStreamT] = AsyncMessageStream,  # type: ignore[assignment]
         # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
         # The extra values given here take precedence over values defined on the client or passed to this method.
@@ -2068,6 +2082,8 @@ class AsyncMessages(AsyncAPIResource):
                     "temperature": temperature,
                     "top_k": top_k,
                     "top_p": top_p,
+                    "tools": tools,
+                    "tool_choice": tool_choice,
                     "stream": True,
                 },
                 message_create_params.MessageCreateParams,
